Make project compile on Linux
- Updated CMake files to support build on Linux
- Deleted original Linux makefiles as they are not needed since CMake is introduced
- Fixed minor problems in the code which prevented the project from compiling

It likely breaks the Windows build, but hopefully it only causes minor problems
those are easy to fix in a way that allows the project to compile on both platforms.

The Linux CMake files don't try to deploy the built libraries to the game mod
directory (people likely build in a separate environment, not where they have the
game installed).