add mp_objectboost cvar to disable object boosting
Clients with high FPS apply their push forces more often, and +USE pushing allows uncapped speeds, so at 200+ FPS you could launch objects/yourself across the map. This cvar disables that, and objects respond to you as if you had 20 FPS or so.

This change also disables player velocity matching with the object, so players don't get towed along if grabbing onto a high velocity object.

Object boosting is fun and useful for speedruns, so this change can be reverted with "mp_objectboost 1"