prevent engine crashing from precache overflow + lowercase resources
errors will be shown in console when a precache limit is exceeded, so you know how close the map is to running, and all the types of resources that need reduction.

All files that are precached are also lowercased to prevent duplicate precache calls with different casing filling up more precache slots (engine precache is case sensitive). This means lowercasing map files on linux is now required.

Bsp-related code copied from bspguy.