fix player gibbing when touching corpses near moving objects
the box was so small that the player entered it instead of standing on top of it, which would make doors and rotating objects think they were being squished between the object and the corpse. The tradeoff for the bigger box is that movement lags when walking over corpses now. Still don't know why things trigger the blocked function when the player is very far away from the object (shattered turbine).