fix StartSound client disconnects
writBit(256) would write a 0 because of the uint8_t parameter type. Also looks like there's a bug in hlds/rehlds where the fractional part of a float is never written because of a misplaced int cast (abs((int)f * 8) & 7). Might want to keep it that way to reduce network usage, or check that rehlds is indeed bugged.

Also re-enabled reload/empty sounds.