add distant gunshots/explosions + performance improvement(?)
gunshots and explosions are heard everywhere in the map now, but are using new distant sounds which are short and quiet because this could potentially be annoying. Ambient noises generally overpower the distant sounds. The purpose of this is to help players find action and let them know that at least one player knows what they're doing. Often times maps will be dead quiet despite a heavy firefight happening less than 1000 units away.

As part of this, all entities need to know which players are in the PAS so that distant sound can be played conditionally based on distance and PAS checks. Knowing this greatly speeds up the UTIL_IsClientInPVS checks that every monster does, which was where most of the cpu time was being spent according to perf. However, the extra AddFullToPack checks may cancel out this speed improvement.