new agrunt hornet firing behavior
If the target is visible, fire faster non-tracking hornets. Use homing hornets only when the target isn't visible. This makes combat with them agrunts less annoying because it gives you a non-zero chance to avoid damage. New behavior is toggled with a skill cvar. Grunts will lead their shots to make it harder to dodge.