expose mod functions for metamod plugins to call
calling the engine version of PrecacheSound was causing crashes due to overflows, because the mod was not aware of the sound being precached by the plugin.

To use these functions, the plugin needs to open the mod dll and call GetModAPI. Then, the functions can be used just like the engine/gamedll funcs in metamod. An example will be added to the mmlib repo.