trigger_entity_iterator speedup
only find the entities to trigger once, if there is a 0 delay between entities. Increases fps about 5x in arcade.bsp.

Finding entities to trigger in the other logic entities is also causing slowdowns because pacman has about 50 changevalues chained together and triggered every frame or so. Not sure what to do about that besides caching targetname search results and refreshing whenever a targetname is changed (complicated and prone to bugs).