prevent edict overflow disconnects
Servers can be configured to use up to 2048 edicts, but clients are disconnected with this error when the server tries sending entity info above the client max.
"Host_Error: CL_EntityNum: 1665 is an invalid number, cl.max_edicts is 1665"

Preventing any entity data or network message from referencing edicts above that limit stops the disconnects, and spams the console instead. An update to rehlds is also needed to stop disconnects when baselines are set when a player first joins.