prioritize edict slots to fix overflows and attachments
this groups the edict list into 3 areas: high, normal, and low priority.

High priority ents are things that can have attachments (e.g. monsters like the garg and it's glowing eye sprite). This area must have indexes lower than 512 because the "skin" keyvalue is sent with 9 bits to the client. Ents with an index above 512 will have broken attachments otherwise.

Low priority ents are things that are never sent to clients. Most triggers and logic entities fit this description; they're silent, invisible, and not used for client-side physics. The low priority indexes are above the default client's max edict count (1665 for HL in a 32 player server).

Normal priority ents are everything else. An entity which will likely be sent to clients for rendering/collision but will not need things attached to it. This is the middle area of the edict list.

In addition to fixing monster attachments, this fixes clients not seeing certain entities because an ent index was too high to be transferred. kh2.bsp for example had a ton of infodecal entities which hogged the lower portion of the edict list, and that caused some important objects like doors and chargers to be invisible. No longer a problem now that the infodecal ents are marked as low priority (one of many server-side-only ents).