add a plugin manager and API generator
This is the replacement for AngelScript. Plugins use the same headers as the mod and can define new entities in the same way. Only a few basic hooks are defined for now. Custom server-side weapons are possible by calling UTIL_RegisterWeapon in a MapInit() hook. Example plugins will be in a separate repo. The cfg command for loading a map plugin is "map_plugin <plugin_file_name>" which loads from the plugins/maps/ game path.