add AddressSanitizer option
This catches memory corruption bugs which are normally hard to debug. It slows down execution speed significantly (google estimates 2x slowdown) so shouldn't be enabled unless investigating a bug.

For example, the last commit had a crash fix for a buffer overrun in a global array. Core dumps were pointing all over the code, then ending in a corrupted frame. Asan pointed right the file+line that caused the overrun. Add "-DASAN=ON" to enable (linux only for now), and recompile rehlds with the same flag or else symbols will be undefined.