add trigger_changesky
sv_skyname is synced to clients but requires existing clients to rejoin in order to see the new sky. So, a bigass model is drawn instead. The conversion script will generate the needed skybox models for the map.

A BSP is used instead of MDL because world lighting doesn't affect BSPs when used as entity models (no "fullbright" flag for MDL textures in HL), and the entity will render everywhere regardless of VIS. There is less color depth, and the linear texture filter mode is not forced for this kind of sky. The max view distance for the map is forced up to a minimum of 2,097,152 for this effect.

Also, the conversion script now handles bad casing on some worldspawn keys and warns about invalid texture sizes on models (no automatic scaling yet).