set edict limits per client
This is done by detecting which version of the game a player is using (possible by querying client cvars). sv_max_client_edicts is still needed engine-side and should be set to the lowest expected max client edicts value (1365 for steam_legacy HL). A warning message is sent to steam_legacy clients when an object/effect was made invisible due to the lower edict limit on steam_legacy.

Also added detection for HLBugFixed and added/fixed some network message utils.