add plugin update commands and cvars
added command "updateplugin" and "updateplugins". The path at cvar "plugin_update_path" is searched for updated plugins when running one of these commands. If a matching plugin in the update path is found, a backup of the current plugin is made, then the updated plugin is moved into place. If the updated plugin fails to load, the backup is restored.

updateplugins will also check plugins.txt for added/removed plugins

if "plugin_auto_update" is "1" then plugins will be updated every map change.