add stukabat
most of the code is copied from the alien controller. It dive bombs targets or flies in close for a melee attack if not high enough. It retreats into the sky after landing a hit. It can be flinched with any caliber higher than 9mm, which cancels its attack. Perching and corpse feeding not implemented.