smooth movement for projectiles and some npcs
Clients don't interpolate velocity on entities that use these movetypes:
NONE, STEP, WALK, FLY

This is true even if you apply the EFLAG_SLERP flag. Only the alien controller and osprey can use that flag because movement is handled without using the engine physics. I don't know why exactly that works. Velocity isn't sent to the client, so it shouldn't know what process changed an entity origin. It should respect EFLAG_SLERP in any case, but it doesn't. Debugging with xash should clear that up.

Other movetypes are interpolated without any hacks, so I'm using those instead. The best interpolated movement type that can fly is BOUNCE. Unlike TOSS, this type can slide along the floor as well as walls.

This smooths movement for all projectiles and a few npcs (notably the apache and ichthyosaur).

To see when the client chooses to interpolate:
https://github.com/FWGS/xash3d-fwgs/blob/51fa7d2611cb2f5da78506ff485a37b1b92bd74b/engine/client/cl_frame.c#L1170-L1264