add kingpin
it has the same abilities as in sven except for the shield. The visual effects are similar but not identical. Most of the sounds are high pitched alien grunt noises. It's still no fun to fight because the lasers are impossible to avoid. The eye laser things have a field of view now, so fewer of them can attack you at once.