add env_weather and env_rain
env_rain is an alias to env_weather. There's no configuration yet and it can only rain.

Rain is displayed as a dense grid of func_conveyors. Droplets are shown on the floor both as a flat texture and particle effects. Rain is dynamic and reacts to breaking and moving objects such as func_breakable, func_train, and func_water. Audio is also dynamic and depends on how many particle effects you can see. A unique sound is played when rain hits a glass ceiling.

No brushes or hints are needed to keep rain out of indoor/roofed areas, but the rain grid is only accurate to 128 units horizontally so there will be gaps where you'd expect rain to be. Rain takes a while to generate when the map loads, but usually less than a second. A cache similar to the node graph cache should be added later.

In the worst case, this adds about 300 wpoly 200 visible entities to an outdoor area. On average it's more like 100 wpoly and 60 entities. It's possible that the entities go outside the valid index range for clients if they're spread out and causing 1000+ ents to be simulated at once. Otherwise, conveyors that no one can see are deleted to keep the edict count low. More model merging can be done later if that becomes a problem (trading precache slots for edict slots).

Network usage is low for the conveyors. The splash particles generate a lot of unreliable data, but less than the data for a single player using voice chat.