add env_fog
The hackiest solution for fog ever. It works using multiple overlapping spheres surrounding the player with different scales and transparencies.

Scale is controlled by animations, and color by submodels which map to a single pixel in a swappable texture. An offset is applied so that the model can be made full bright with the EF_BRIGHTLIGHT effect, without the light being visible.

It looks good with 64 layers, but that's too much entity data to send. These can't be attached to players with aiment because then the animations break. So, it uses 16 layers for now. It looks very retro, even for HL.

Transparent entities are bugged with fog. Anything in the fog volume is made invisible unless changing rendermode back to normal. This will happen for everything except beams. Sprites mostly look correct. Windows will suddenly become solid sometimes, especially if multiple are tied together to form a large bounding box.