fix clients freezing in areas with lots of entities
when a client enters a new area with 200+ ents, it freezes with message "datagram overflowed" spammed in the server console. Sometimes the client eventually receives all the data, but not always.

Another source of freezing comes from MAX_PHYSENTS. If near the limit, the player will become completely stuck. Not even noclip can save them.

This change limits the amount of new entities per frame so that players gradually receive the full state. Some entities will be invisible while catching up, which is less than a second at cl_updaterate 10, and almost instant at the default 60. MAX_PHYSENTS was also increased from 600 -> 2048. A rehlds update is required or else the server crashes instantly.

MAX_PACKET_ENTITIES was lowered to a practical limit. The client can't actually handle 1024, it's more like 500.