add option to build a server lib with increased limits
the MAX_PHYSENTS change breaks compatibility with the client engine. In order to raise this limit for a dedicated server, both rehlds and the server library need to be compiled with -DINCREASED_SERVER_LIMITS.

Increased MAX_PHYSENTS prevents crashes and players freezing in areas with lots of entities (cow_tycoon fences or rust without build limits).