fog fixes
The entire fog volume flickers on/off when shooting which is very distracting. The problem is caused by the muzzle flashes created by the client engine.

Adding an offset to the model fixes this, but the minimum fog distance needs to be increased again to prevent rendering glitches. It's now 450 for black fog, and 650 for colored fog. Previously it was 384 for both.