add mp_series_intermission
if this set to 1, game_end functions like trigger_changelevel for maps in a series. If set to 2, trigger_changelevel and series game_ends will briefly show the scoreboard and a level change message in chat (MapChangeMsg plugin behavior).

The point of mode 2 is to let players know why the server is frozen. It's not a server crash. Sometimes the server takes a while to change levels (node graph generating for example). Most times players probably won't be able to read the level change message because it switches so quickly.