integrate AntiBlock and tweak +use behavior (#44)
https://github.com/wootguy/AntiBlock
This plugin is essential and belongs in the mod code. Enable with mp_antiblock 1.

Hold your use key on a player or ally npc to swap places with it. When antiblock is enabled, entities aren't activated until the use key is released, so that you don't accidentally activate an npc while trying to swap with it. Holding the use key longer than 0.3s cancels the use attempt.

mp_antiblock_cooldown controls the cooldown when doing a rude swap. A rude swap is when your target is a player that isn't looking at you. Probably you are being impatient in a vent or trying to swap spam to annoy someone. The cooldown doesn't apply if the player has been idle for 5+ seconds.

Use behavior was tweaked so that the entity you're looking directly at has the highest priority, if close enough. This fixes the problem where there are multiple buttons in a small area and somehow a button you're not looking at is activated.