allow weapons to share menu slots
A hacky solution that involves lying to the client. When a player picks up a weapon that shares a slot with other weapons, the client will be told that they have every possible weapon for the slot. User messages then overwrite the client's knowledge of other weapons that can use the slot, preventing missing/wrong icons.

If a player attempts to hold multiple weapons that share a slot, the held weapon will be dropped to make room for the newly collected one.

also precaching custom weapons used in the map cfg.