improve lag compensation performance
no need to check entities that can't be networked to the client. Also don't need to call SetOrigin if the origin hasn't changed. That's the most expensive part of this according to perf. On a map like alamo1 this should run about 20x faster (average of 20 SetOrigin calls per call down from ~400).