large map support + fixes/tweaks
A collection of hacks to make areas outside +/-4096 more playable. Most visual and audible effects didn't work past this boundary because they use network messages with limited coordinate ranges. By using normal entities for the effects instead, the max range of these effects is increased to whatever you set in delta.lst.

This dramatically increases network usage to the point that it will likely overflow and disconnect players if there's a lot of action outside +/-4096. The effects are also strange and can't match the originals exactly. Many are not interpolated and spawn fewer particles. Player sprays still don't work at all, and other decals won't wrap around objects. Not all effects are implemented yet. Fighting NPCs feels mostly normal.

Set "mp_bigmap 1" in the map cfg to precache the needed models/sounds for playing outside the +/-4096 HL limits. For a small area with maybe 10-20 npcs this should be fine, and save a lot of time trying to split a map to fit inside the HL range. For instance, sc_castlevania would be a nightmare or impossible to fit into the grid well enough to prevent glitches. It doesn't have much action in it though so I'd rather deal with the side effects of mp_bigmap.

Fixed:
- wildcard searches finding string that didn't start with the search string
- missing newlines on error messages
- client crash setting non-sprite on beam (e.g. the not precached model)
- missing code from previous commit for removing high priority edict slots
- sprites animate if not given a framerate (fixes alien controller orbs)

Tweaks:
- env_sprite can play up to 20fps, up from 10
- added helpers for getting a model name from an index