cstrike crossplay wip
clients can join the server with a crossplay plugin or if sv_lan is enabled, but the server player count doesn't go up because the steam ticket is invalid for the server's app id. That can be worked around by creating dummy steam accounts that generate tickets for HL in place of real player tickets. 32 instances of the steam client would need to be sandboxed and running on a server somewhere. Maybe a sketchy thing to do. I couldn't get the steam client to run on my headless arm vps after a few days, so giving up for now. Looks like goldsrc.one doesn't do this. Players from other games are counted as bots which is why the server looked super dead on trackers.