partially fix stuttering voice audio
happens reliably at the start/end of cutscenes. The "fix" sends a stopsound command to the client and restarts ambient sounds. This also sometimes happens when moving around or teleporting so sound is reset on spawning too. It's now much more important that map music plays using the mp3 command.