support hybrid HL/SC player models
add support for a new type of player model which combines HL and SC animations in a backwards compatible way. New Sven Co-op 5.x animations are moved after the end of the default Half-Life animations. This way, models can be used on vanilla Half-Life servers without broken animations (dying upright, swinging guns like a crowbar, etc.).

The server needs to have all used player models installed so that it knows which model names use the new animations. There is no way yet to check if a player's model file matches the server, but this should be done since there are tons of models that share the same name and are completely different.

hybrid models have new exclusive animations:
- unique reload/shoot/deploy for every weapon
- charge-up for the wrench and grenade
- barnacle victim capturing and head chewing

HL model logic was updated to be consistent with SC:
- arms move when running/walking/crawling without weapons
- more recoil for shockrifle, grapple, displacer, and mp5 nades
- wrench/grenade is held lower so charge-up animation can be shown
- the grapple is held like the hornetgun instead of shotgun
- added a barnacle victim pull animation
- xbow doesn't show a muzzle flash when firing from the hip
- xbow has a zooming animation