custom predicted weapon prototype
Custom weapons with prediction are implemented as a set of parameters for controlling which animations/sounds/etc. to play when using the weapon. This allows new weapons to be created without touching client code. A scripting system wasn't used because most weapons don't need that complexity. The rare complex cases are usually projectile firing weapons which can't be predicted anyway. CWeaponCustom implements basic weapon functionality that all children weapons share.

Prediction parameters are sent as a network message when picking up the weapon. Included in the parameters is a list of events for client-side effects. These are things like bullet tracers, sounds, and animations. They can be attached to any weapon action with a custom delay.

A separate message is sent when the player joins the server which maps sound indexes to sound paths. This is needed because the client can only play sounds by name, not by index. Sound paths would take up too much space if added to the prediction data.

The M249 is now implemented as a CWeaponCustom child. It only works when using the custom client so the mp5 replacement is still active.

Also fixed a bug on a client side that prevented weapon deploy animations from playing.