rework thirdperson camera and allow multiplayer use
a new +thirdperson command can be pressed to toggle thirdperson, or held down to rotate and zoom the camera. Most of the cam_* cvars were removed in favor of this, and cam_snapto was replaced with cam_snapback, which controls the camera resetting to the default rotation/distance a short time after -thirdperson.

+thirdperson is only allowed on sevenkewp servers because it can be used to cheat in deathmatch. SevenKewp servers send a special network message signaling to the client that it's ok to use the command.