custom hud.txt update server info, increase weapon slots
weapon slots were increased to 6 and the M249 was moved to that slot. The ServerCfg user message was removed in favor of the serverinfo string, which can be parsed by the client before they fully join. This allows a custom hud.txt path to be loaded during the connection phase. A makeshift client-side hook detects when all files have downloaded and the player is about to spawn.