auto-update hud file
the server and client hash this file. If the client hash doesn't match, then it deletes its local copy of sprites/hlcoop/hud.txt and redownloads from the server. Only files in valve_downloads/ are deleted. An override in valve_addon will trigger a mismatch on every connect but still allow joining.