add m16 from sc
Added for the sake of completeness, but I'm not adding it to the game.

Most of the projectile logic was copied over from the weapon_custom plugin to CWeaponCustom. Much of it isn't implemented.

TODO:
The secondary clip/reload isn't implemented. The burst fire logic needs improvement (decrement ammo on each shot, not 3 at once). Shoot events are failing to fire sometimes (all custom weapons probably).