build compatible client binaries
clients that can run on an OS as old as Windows 7 or Ubuntu 18.04 without any setup beyond copying the files.

For Linux, I wanted to target 12.04 (first version to support Steam) but it looks like SDL dependencies have breaking changes every few years. I couldn't even get Steam running on 12.04 or 14.04 anyway. A build for 18.04 was as old as I could go before the library failed to load on 24.04.

For Windows, I wanted to target XP but installing the v141_xp toolset takes 5 minutes on the github hosted runner. I tried creating a docker image with that preinstalled, but the image takes 5 minutes to extract. I gave up after trying and failing to reduce the image size. The client can't actually run on XP yet anyway because the HL25 update broke support for it, and the client crashes with steam_legacy. Windows 7 was targetted instead. The auto-updater fails on 7 due to TLS but I read that's something you can patch the OS to fix. The alternative would be including certificates with the .dll.

I tried to get a mingw build working but it fails to link because of the c++ dll interface in VGUI. It's a big job to work around that.

also fixed client crash with increased max weapons (forgot to commit rehlds change)