add a command menu and move the client update command there
custom client console commands aren't protected by cl_filterstuffcmd so the command menu is used to be sure that the client initiated the update or game restart.

commandmenu_sk.txt is included in the auto-updated file list. This might be a problem later because players sometimes customize this menu. There should be a multi-file menu separating player, server, and mod commands.

A bunch of TFC code was removed too.