chat tweaks
hud_saytext_time now sets the display time for each message individually, not as a delay for scrolling up the history. The old behavior displayed messages much longer when multiple were on screen at once (up to 25 seconds for 5 messages). The default value of hud_saytext_time is doubled from 5 to 10 to compensate.

also:
- chat history is shown when typing a message
- chat grows from bottom to top (less likely to block view)
- fixed chat input displaying at the top of screen for the first message