add mp_keep_inventory
Works the same as the keep_inventory keyvalue for trigger_changelevel when set to 1. Kept inventory is added in addition to the map default, and health/armor is restored.

When set to 2, the exact state of the player is saved and loaded in the next map. That includes health, armor, active weapon, flashlight battery, and weapon clips. If the player ever dies, their equipment is reset to the map default.